Rebase and resolve comments
